requirements.txt
fastapi>=0.111.0,<0.115.0
uvicorn[standard]>=0.23.0,<1.0.0
jinja2>=3.1.0,<4.0.0
PyYAML>=6.0,<7.0
python-docx>=0.9.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
aiofiles>=23.2.0,<24.0.0
pydantic>=2.7.0,<3.0.0
